add fastapi service for model
