+++ requirements.txt
@@ -4,3 +4,6 @@ openai~=1.65.2
 langchain~=0.3.20
 python-dotenv~=1.0.1
 pydantic~=2.10.6
+psycopg2~=2.9.10
+uvicorn~=0.34.2
+fastapi~=0.115.12
